Added App icon, iPad support, and refine Physics Body
The App now has an icon in 2x and 3x resolution, suitable for iPhone
and iPad. To support iPad, the screen orientation is limited to
portrait only. Centered and repositioned the Storyboard with Loading
message to adjust for any screen size. Refined Physics body for some
objects to better match the shape of the sprite.